api/requirements.txt
-i https://pypi.org/simple
absl-py==0.7.1
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
flask==1.0.3
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
importlib-metadata==0.17
ipykernel==5.1.1
ipython-genutils==0.2.0
ipython==7.5.0 ; python_version >= '3.3'
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter==1.0.0
keras-applications==1.0.8
keras-preprocessing==1.1.0
kiwisolver==1.1.0
markdown==3.1.1
markupsafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0 ; python_version > '2.7'
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
packaging==19.0
pafy==0.5.4
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pillow==6.0.0
pluggy==0.12.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0 ; os_name != 'nt'
py==1.8.0
pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.6.2
python-dateutil==2.8.0
pyzmq==18.0.1
qtconsole==4.5.1
send2trash==1.5.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
werkzeug==0.15.4
wheel==0.33.4 ; python_version >= '3'
widgetsnbextension==3.4.2
youtube-dl==2019.5.20
zipp==0.5.1
